Common middleware for Angular, React and other dev servers that can be run by npm command
